refactor(sync): isolate database synchronization pattern to repository tier and harden with Promise.allSettled and Zod
